# api/openapi.yaml
openapi: 3.1.0
info:
  title: Alexandria User Service API
  description: >-
    User account management. Owns the users table; exposes internal fan-out
    endpoints consumed by knowledgebase-service and qa-service.
  license:
    name: MIT
    identifier: MIT
  version: 2.0.0
servers:
  - url: /
    description: Current server
...
paths:
...
  /api/v1/users/{id}:
    delete:
      tags:
        - User Service
      summary: Delete user account
      description: Users can only delete their own account
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted
        '403':
          description: Cannot delete other users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
...
  /internal/knowledgebase/users/{subject}/document-keys:
    get:
      tags:
        - KnowledgeBase Service (Internal)
      summary: List object keys for a user's documents (internal)
      operationId: listDocumentKeys
      parameters:
        - name: subject
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object keys returned
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      security: []
...
  /internal/knowledgebase/users/{subject}:
    delete:
      tags:
        - KnowledgeBase Service (Internal)
      summary: Purge all knowledgebase data for a user (internal)
      operationId: kbInternalDeleteUserData
      parameters:
        - name: subject
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User data purged
      security: []
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...